python/requirements.txt
autopep8==2.0.0
docopt==0.6.2
hypothesis==6.60.0
ipython==8.7.0
pylint==2.15.9
PyYAML==6.0
requests==2.28.1
tabulate==0.9.0
